refactor(mediatask): seperate snapshot & subtitle
Update electron and seperate SnapshotQueue and SubtitleQueue.

1. Update electron version to 5.0.7.
2. Seperate snapshot and subtitle to different queues.
