+++ package.json
@@ -105,7 +105,7 @@
     "@babel/plugin-syntax-import-meta": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
     "@babel/register": "^7.0.0",
-    "@chiflix/electron": "5.0.6",
+    "@chiflix/electron": "^5.0.7",
     "@sentry/webpack-plugin": "^1.6.2",
     "@types/chai": "^4.1.7",
     "@types/lodash": "^4.14.134",
